removed unnescessary casting of integer value

+++ adm_program/modules/dates/dates.php
@@ -33,8 +33,8 @@ require_once(__DIR__ . '/../../system/common.php');
 unset($_SESSION['dates_request']);
 
 // Initialize and check the parameters
-$disableAdditionalGuests   = (int) 2;
-$disableComments           = (int) 2;
+$disableAdditionalGuests   = 2;
+$disableComments           = 2;
 $disableStatusAttend       = '';
 $disableStatusTentative    = '';
 $getMode     = admFuncVariableIsValid($_GET, 'mode',      'string', array('defaultValue' => 'actual', 'validValues' => array('actual', 'old', 'all')));
@@ -615,8 +615,8 @@ else
                         </div>');
                 }
                 // Reset flags and parameters
-                $disableAdditionalGuests    = (int) 2;
-                $disableComments            = (int) 2;
+                $disableAdditionalGuests    = 2;
+                $disableComments            = 2;
                 $disableStatusAttend        = '';
                 $disableStatusTentative     = '';
                 $participateModalForm       = false;
